LPS-107004 Break into Assignee and AssigneeMetric
