+++ modules/dxp/apps/portal-workflow/portal-workflow-metrics-rest-impl/src/main/java/com/liferay/portal/workflow/metrics/rest/internal/odata/entity/v1_0/AssigneeMetricEntityModel.java
@@ -28,9 +28,9 @@ import java.util.stream.Stream;
 /**
  * @author Rafael Praxedes
  */
-public class AssigneeEntityModel implements EntityModel {
+public class AssigneeMetricEntityModel implements EntityModel {
 
-	public AssigneeEntityModel() {
+	public AssigneeMetricEntityModel() {
 		_entityFieldsMap = Stream.of(
 			new IntegerEntityField(
 				"durationTaskAvg", locale -> "durationTaskAvg"),
@@ -52,7 +52,7 @@ public class AssigneeEntityModel implements EntityModel {
 	@Override
 	public String getName() {
 		return StringUtil.replace(
-			AssigneeEntityModel.class.getName(), CharPool.PERIOD,
+			AssigneeMetricEntityModel.class.getName(), CharPool.PERIOD,
 			CharPool.UNDERLINE);
 	}
